Adding collapsible UI component on Dashboard

+++ food-donator-frontend/src/app/components/donor-dashboard/donor-dashboard.component.ts
@@ -3,6 +3,7 @@ import { Donation } from 'src/app/models/donation.model';
 import { AuthenticationService } from 'src/app/services/authentication/authentication.service';
 import { DonationService } from 'src/app/services/donation/donation.service';
 import { SidenavService } from 'src/app/services/sidenav/sidenav.service';
+import { UserTagService } from 'src/app/services/user-tag/user-tag.service';
 import { Constants } from 'src/app/shared/constants/constants';
 import DateUtil from 'src/app/utils/DateUtil';
 
@@ -12,34 +13,57 @@ import DateUtil from 'src/app/utils/DateUtil';
   styleUrls: ['./donor-dashboard.component.css']
 })
 export class DonorDashboardComponent {
-
   currentAndUpcomingDonations: Donation[] = [];
   pastDonations: Donation[] = [];
+  currentDonationCounter = 0;
 
-  constructor(private donationService: DonationService,
-    private authenticationService: AuthenticationService,
+  constructor(
     public dateUtil: DateUtil,
-    public sidenavService: SidenavService){
+    public sidenavService: SidenavService,
+    public donationService: DonationService,
+    private authenticationService: AuthenticationService,
+    private userTagService: UserTagService
+  ) {
     const jwt = window.sessionStorage.getItem(Constants.FOOD_DONATOR_TOKEN);
     this.authenticationService.getUserByJWT(jwt).then((user) => {
-      if(user != null) {
+      if (user != null) {
         this.donationService.getCurrentAndUpcomingDonationsByUserId(user.id).then((donations) => {
-          if(donations != null) {
+          if (donations != null) {
             this.currentAndUpcomingDonations = donations;
           }
         });
 
         this.donationService.getPastDonationsByUserId(user.id).then((donations) => {
-          if(donations != null) {
+          if (donations != null) {
             this.pastDonations = donations;
           }
         });
       }
     });
+
+    $(() => {
+      // initialize the current and upcoming donations collapsible UI component
+      $('#currentCollapsible').collapsible({
+        accordion: false
+      });
+
+      // initialize the past donations collapsible UI component
+      $('#pastCollapsible').collapsible({
+        accordion: false
+      });
+    });
   }
 
-  async deleteDonation(donationId: string) {
+  async deleteDonation(donationId: string, e: Event) {
+    e.stopPropagation();
+
     try {
+
+      //TODO replace with modal popup
+      if (!confirm('Are you sure you want to delete this donation?')) {
+        return;
+      }
+
       await this.donationService.deleteDonation(donationId);
       this.currentAndUpcomingDonations = this.currentAndUpcomingDonations.filter((donation) => {
         return donation.id != donationId;
@@ -48,10 +72,9 @@ export class DonorDashboardComponent {
       this.pastDonations = this.pastDonations.filter((donation) => {
         return donation.id != donationId;
       });
-    } catch(e) {
+    } catch (e) {
       console.error("Could not delete donation: ", e);
     }
 
   }
-
 }
